archive/AdvancedAttendance/requirements.txt
insightface==0.7.3
torch>=2.0.0
opencv-python>=4.5.0
PyQt5>=5.15.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-telegram-bot>=20.0
pygame>=2.0.0
loguru>=0.6.0
numpy==1.24.3
onnxruntime==1.15.1
pyyaml>=6.0
jinja2>=3.1.0
requests>=2.31.0
websockets>=11.0
sqlalchemy>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.1.0
